# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Patient Price Discovery API
  description: API for searching and viewing healthcare facilities and prices.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/facilities:
    get:
      summary: List facilities
      description: Retrieve a list of facilities with optional filtering.
      parameters:
        - name: type
          in: query
          description: Filter by facility type (e.g., "Hospital", "Imaging Center")
          schema:
            type: string
      responses:
        '200':
          description: A list of facilities
          content:
            application/json:
              schema:
                type: object
                properties:
                  facilities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Facility'
                  count:
                    type: integer
                    description: Total number of facilities returned
        '500':
          description: Internal server error
...
  /api/facilities/{id}:
    get:
      summary: Get facility by ID
      description: Retrieve detailed information about a specific facility.
      parameters:
        - name: id
          in: path
          description: ID of the facility to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Facility details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Facility'
        '404':
          description: Facility not found
        '500':
          description: Internal server error
...
components:
  schemas:
    Facility:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        location:
          $ref: '#/components/schemas/Location'
        phone_number:
          type: string
        email:
          type: string
          format: email
        website:
          type: string
          format: uri
        description:
          type: string
        facility_type:
          type: string
        accepted_insurance:
          type: array
          items:
            type: string
        rating:
          type: number
          format: float
        review_count:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - address
        - location
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zip_code:
          type: string
        country:
          type: string
    Location:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
...